phy: freescale: imx8mq-usb: add support for 8MP/8MQ vbus-supply
The PHY device tree node may reference a regulator via vbus-supply that
needs to be switched by software.

Linux already supported this binding, so implement it for barebox as well.

Unlike the Linux driver, we do not treat a missing regulator as hard
error for backwards compatibility reasons. Instead, we just print a
warning and move along:

- If we have a driver, then on deep probe systems, this should always
  succeed anyway.

- If we do not have a driver and are on deep probe system, the VBUS is
  powered on already, then the warning can be suppressed by
  barebox,allow-dummy-supply in the provider node.

- If the user is not on a deep probe system, they should either patch
  out vbus-supply or enable deep probe.

Reported-by: Holger Assmann <h.assmann@pengutronix.de>
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251103151935.2368094-2-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>