fs: nfs: do not read past packets
The NFS code has hardly any length checks and it blindly trusts length
values in incoming packets. Fix this by introducing a nfs_packet_read()
which is consistently used to read from an incoming packet.

After a packet has been consumed it must be freed with
nfs_free_packet(). This wasn't done in several error cases and is fixed
here as well.

Link: https://lore.barebox.org/20251104-nfs-v1-2-3e69e5c82a41@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>