ARM: k3: allow to take signing key from environment variable
This adds a new Kconfig option CONFIG_ARCH_K3_SIGNING_KEY_ENV. With
this option enabled the content of CONFIG_ARCH_K3_SIGNING_KEY is treated
as the name of an environment variable from which the private key
file/uri is taken. This allows for better integration with build
systems.

Link: https://lore.barebox.org/20251106073035.3580913-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>