.clangd: remove -enable-trivial-auto-var-init-zero-... flag
Debian 12 sill had clang-12, where specifying the option:
-enable-trivial-auto-var-init-zero-knowing-it-will-be-removed-from-clang
was needed to avoid a warning when the build was configured for GCC with
-ftrivial-auto-var-init=zero.

Since clang-16, this option has been removed and it now results in a warning
instead of avoiding one.

Given that Debian 13 has clang-19, let's just drop the flag to remove
the warning for developers with a recent clang version.

This reverts commit 2c151ad3fb19 ("kbuild: fix clangd warning with GCC
compile_commands.json").

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251106143640.224768-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>