crypto: Use "development" keys for "fit" and "tlv" keyring
All users of the CONFIG_CRYPTO_PUBLIC_KEYS feature should update to the
new syntax making keyring selection mandatory.

Instead of just making the addition of the builtin snakeoil keys
explicit for the "fit" key, also add them to the "tlv" key to use them
as a testing set for TLV keys too.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.barebox.org/20251106-tlv-signature-v3-12-5d00ed378e75@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>