Documentation: migration-2025.12.0: List changes to CONFIG_CRYPTO_PUBLIC_KEYS
As part of the introduction of TLV signature in public key keyrings,
there is a new interpreter for CONFIG_CRYPTO_PUBLIC_KEYS in
scripts/keytoc.c. All users should update to the new syntax and are
presented with a warning. Users with invalid characters in the legacy
syntax must change the fit-hint and will be presented with an error.

Signed-off-by: Jonas Rebmann <jre@pengutronix.de>
Link: https://lore.barebox.org/20251106-tlv-signature-v3-17-5d00ed378e75@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>