bobject: free object name in bobject_del()
We have bobject_free() which removes all parameters and frees the
bobject. This works for bobjects which have been allocated with
bobject_alloc(). struct device has a bobject embedded and is freed
along with the device, so bobject_free() is inappropriate and
bobject_del() is called instead. This however doesn't free the bobjects
name which then leaks when the device is freed.

To fix this move the freeing of the bobject name to bobject_del(). To
make this work with CONFIG_PARAMETER disabled drop the static inline
wrapper. As the bobject name is now freed in bobject_del() we have to
remove the freeing of struct device::name which is the same object.

While at it fix the function name in the bobject_del() function
description.

Link: https://lore.barebox.org/20251107-memleaks-v1-3-c7b74a2ae9db@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>