hush: fix memory leaks
The memory leaks in hush have been bothered me for years and every time
I look into it, I give up after a while, because the code is convoluted.

Talloc shines here as we can just allocate a "zero-size" talloc object
as parent for a context and associate all allocations from that context
with it.

Then when that context is free'd all children will be freed as well.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251027075438.2480311-3-a.fatoum@barebox.org
Link: https://lore.barebox.org/20251107-talloc-v2-3-e47bfd0e5667@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>