video: stm: make multi-image compatible
Unlike newer platforms, it's not possible to mix i.MX23 and i.MX28 SoC
support in the same build. The stmfb driver is used on newer SoCs
however like the i.MX6SX and i.MX8MQ/M/N, so remove the #ifdeffery in
the driver.

While at it, we also drop unused macros as well as the unused dotclk_delay.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251107210033.2229781-1-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>