soc: imx8mp: Soc ID is 128bit
On i.MX8MP the SoC ID has 128 bits instead of 64 bits as on other i.MX8M
SoCs. Read the remaining 64 bits which so far haven't been included in
the SoC ID.

On already rolled out devices a change of the SoC ID is undesired, so
this commit introduces CONFIG_ARCH_IMX8MP_KEEP_COMPATIBLE_SOC_UID. With
this option enabled the old SoC ID will be used.

Link: https://lore.barebox.org/20251117-soc-uid-v2-2-a2415bf9133d@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>