nvmem: k3: add fuse support
This driver reads/writes to the extended OTP area using TF-A calls.

The driver has been tested on AM625 and AM62L, but should work on other
SoCs as well. The driver needs the SIP calls K3_SIP_OTP_READ and
K3_SIP_OTP_WRITE. These are currently only implemented in the TI
downstream TF-A for AM62L. For AM625 these calls need to be enabled with
additional TF-A patches. The driver is activated in the AM62L/AM625
device trees, but due to the limited availability of the SIP calls
the driver remains silent when the TF-A doesn't have support for
manipulating fuses.

Link: https://lore.barebox.org/20251125114007.1198441-1-s.hauer@pengutronix.de
Link: https://lore.barebox.org/20251126141954.2033860-1-s.hauer@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>