environment: prefer non-removable over removable devices for autoprobe
On systems with only 1-2 MiB of boot flash, it can be sensible to locate
the barebox environment outside of the boot medium.

In that case, it's sensible to give preference to non-removable devices,
so e.g. an eMMC is used instead of the SD-Card when booting from
neither, so adapt the scoring accordingly.

This could be further refined to make use of the eMMC non_removable
attribute, but this is already arguably an improvement.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251127124140.2229718-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>