kbuild: drop non-existent targets from CLEAN_FILES
A number of these targets are a no longer generated and if support for
them were to be added, it would be PBL-prefixed and handled in images/,
so these left-overs can safely be dropped.

Nothing needs to be removed in .gitignore, as a /barebox* rule is
matching all of the removed files and more.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251128104143.1578832-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>