ARM: mmu: set up vectors prior to full MMU setup
Interleaving vector table setup with zero page setup introduces an
unfortunate dependency: We need to call setup_pages after mapping the
SDRAM, so mapping the zero page as faulting is not overridden and we
need to do it before mapping the text area read-only, because the
vectors are relocated on ARM32.

Avoid this issue by setting up the vectors at core_initcall and only do
trap page setup during MMU init. This is already what is being done on
arm64.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251128172206.36551-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>