fs: split rootargs into root and options
The rootargs argument (root=/dev/etc rootwait) can be split into the
root and options parts, which makes it easier to manipulate them both
independently. This changes tries to be as backward compatible as
possible and does not change the behaviour of the get functions, but
introduces new ones to get both items seperatly.

Signed-off-by: Fabian Pflug <f.pflug@pengutronix.de>
Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251201072452.1829564-3-f.pflug@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>