of: add proper driver for simple-pm-bus
simple-pm-bus is currently handled exactly the same as simple-bus in
barebox with the assumption that clocks/power-domains are already
enabled.

This is not necessarily the case, so let's add a driver for
simple-pm-bus that actually does power management and keep the old
behavior the same for devices without deep probe.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251201084737.1270303-4-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>