pmdomain: genpd_activate() automatically for deep probe platforms
With the addition of the SCMI power domain driver that's applicable to
any ARM SoC, it's not appropriate to keep it a SoC-specific decision,
whether power domains need to be handled.

That scheme also breaks down as soon as we have some power domains that
need to be handled and some that don't.

The original commit justifies the addition by support of OMAP HW in
multi_v7_defconfig. We do not support deep probe on OMAP yet, so
it's unaffected and when we do, barebox,allow-dummy can be peppered
around the device tree to get rid of the warning.

If anybody else gets warnings from this, they need to add
barebox,allow-dummy to the single power domain in question.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251201084737.1270303-6-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>