video: dw_mipi_dsi: activate pattern generator during enable as a hack
There's still something fishy with the initialization sequence of the
video pipeline with respect to the MIPI-DSI bridge.

As a workaround, jump start the pipeline by briefly enabling the video
pattern generator and disabling it again.

This will lead to a very brief flash of colored bars until that issue is
better understood and properly resolved.

Signed-off-by: Ahmad Fatoum <a.fatoum@barebox.org>
Link: https://lore.barebox.org/20251201085927.1306364-3-a.fatoum@barebox.org
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>