pmdomain: fix NULL pointer derference with dummy power domains
With the addition of commit 4aca68990d04 ("drivers: power: ignore power
domains with barebox,allow-dummy"), it's possible for genpd to be NULL
when there is no driver and the device tree node has barebox,allow-dummy
in it.

Allow for this possibility in genpd_add_device and genpd_remove_device
as well.

Fixes: 44c6bedbd439 ("pmdomain: allow callback for when devices are attached")
Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251201133832.3303902-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>