partitions: gpt: write inactive GPT first
There are two GPT partition tables on a device, the primary at the
beginning of a device and the alternate at the end of the device.

For the sake of being more robust against power failures let's write the
inactive GPT first and the currently active one afterwards. This always
leaves us a usable GPT on the device when updating the partition table.

Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251203-efi-partition-refresh-v1-4-f0b6e79b5fa0@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>