nvmem: fix device name setting
For the nvmem device ID it was only possible to specify NVMEM_DEVID_NONE
which translates to DEVICE_ID_SINGLE or NVMEM_DEVID_AUTO which
translates to DEVICE_ID_DYNAMIC. All other values default to
DEVICE_ID_DYNAMIC.

Explicitly allow to set a fixed device ID so that the caller can fully
specify the name under which the new nvmem device appears. The special
case is when no config->name is given. In this we use "nvmem" as device
name, but we also have to ignore config->id.

Link: https://lore.barebox.org/20251208-nvmem-eeprom-v1-2-41ba588e2f3a@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>