storage-by-alias: drop fake cdev
storage-by-alias creates its own cdev, but that is never used. The only
purpose is to create a cdev with its own device node which otherwise
links to the original cdev. This means we can drop all the cdev
operations and use the newly created devfs_create_link_node() with the
same effect.

Reviewed-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251209-devfs-v2-2-62ae16698cff@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>