fs: replace cdev links with aliases
cdev links are cdevs of their own, so we have to add cdev_read_link() in
several places to get to the original cdev.

Replace this with cdev aliases. Aliases are no cdevs themselves, but a
list of aliases inside the original cdev which gets us a more straight
forward implementation.

Acked-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251209-devfs-v2-12-62ae16698cff@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>