clk: vexpress: clk-sp810: sync with Linux
labgrid-pytest --lg-env test/arm/a15@multi_v7_defconfig.yaml crashes for
me, because the parent is hardcoded as 1 and that leads to an out of
bounds access when executing the clk_dump commands as it tries to get
ahold of the name in clk_lookup.

Sync the driver with Linux to fix this issue.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251211174050.3689753-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>