ARM64: support reserving the x18 platform register
Linux helpfully complained about "register x18 corrupted by EFI" with
the upcoming EFI loader support as this register is apparently the
platform register and is meant to be usable by the OS without other
firmware like the EFI runtime services clobbering it.

Provide a hidden AAPCS64 symbol that will be usable for this purpose.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251215084711.2635214-1-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>