resource: assign memory banks a default type and attr
The resource_iter family of functions iterates over both resources and
the gaps between them and will inherit the parents type and attrs into
the gaps.

Give the SDRAM suitable attributes for the gaps, so users need not
special case the region_is_gap case if they are interested in collecting
all memory type and attributes.

Link: https://lore.barebox.org/20251218111242.1527495-9-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>