efi: loader: add object handling API
Add definitions and prototypes for the EFI object model.
The implementations will follow later, but we add this first here, so we
do not have to add everything in one big commit and can split it out
into separate parts.

The code using it can still be compiled, but it will only be linkable
with linker garbage collection enabled.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251218111242.1527495-16-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>