efi: loader: add devicepath support
A proper device path would walk the chain of barebox devices as
determined from the OF device tree and assemble a path like:

  MMIO -> PCI root -> USB controller -> USB Mass storage -> Block device

This can still happen in future, but for now we just use following
simplification that allows us to represent every VFS path in barebox as
EFI device path:

  - A Bbs device node for the device component:
    * The cdev of the device the file resides in
    * The mount path followed by / if no cdev backs the FS device
  - A FilePath device node for the file path, relative to the mount point

BBS is the BIOS Boot Specification, which seems long obsolete, but it lends
itself nicely to our purposes of wrapping barebox VFS paths. There's already
a FilePath device node sub type for media devices, but it's usually at the
end of the device path and buggy EFI payload may be not search beyond the
first file path they encounter. All other device path nodes are documented
as formatting their argument as hexadecimal with BBS being the only exception,
which takes a non-UTF-8 string.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251218111242.1527495-17-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>