efi: loader: protocol: add efi_device_path_to_text support
For use by EFI payloads that don't want to convert device paths into
strings themselves, the loader can provide a protocol with access to its
own routines. device_path_to_str(), which we already in the EFI payload
can fill that role, so add a protocol that calls into it.

Signed-off-by: Ahmad Fatoum <a.fatoum@pengutronix.de>
Link: https://lore.barebox.org/20251218111242.1527495-36-a.fatoum@pengutronix.de
Signed-off-by: Sascha Hauer <s.hauer@pengutronix.de>